Mock searches against no indexes
Rather than rely on elasticsearch itself to give us an empty msearch
result, we outright mock the empty result in kibana and avoid making an
msearch request altogether.

While relying on elasticsearch is ideal, the approach is flawed because
the only way to guarantee that a search queries no indexes and returns
no results is to query against -*, but such a query will never authorize
successfully against shield.

This does introduce a change to the public api of the courier fetch
strategies, where each strategy is now responsible for executing a
provided function to retrieve the results. This allows each strategy to
optionally decorate the fetching logic as necessary.

For now, this change does not break backwards compatibility as the
existing fetching behavior is preserved for any strategy that does not
provide its own execute method, but for the sake of explicitness, that
fallback behavior should be considered deprecated from this point
onward.