server: fail to start with unrecognized configs
Having an unrecognized configuration could result in a dangerous
deployment scenario if it happens as a result of a misunderstanding or a
typo, so it is safer for production installs to not even start if an
unrecognized configuration is present. For convenience sake, we don't
kill the server when in development and just log a message instead.